pyproject.toml
[tool.poetry]
name = "graph_traversal"
version = "0.1.0"
description = ""
authors = ["Fred O'Loughlin <fred@oloughl.in>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.1.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
